[2018 spring rework] Correction de bugs : - ChunkKeys::toString - vec3i/vec3d operators - segfault à la fermeture de world3D

+++ modules/worldapi/core/WorldKeys.cpp
@@ -10,7 +10,7 @@ namespace world {
     std::string ChunkKeys::toString(const world::ChunkKey &key) {
         std::stringstream stream;
         for (char c : key) {
-            stream << std::setfill('0') << std::setw(2) << std::hex << (u8) c;
+            stream << std::setfill('0') << std::setw(2) << std::hex << (int) c;
         }
         return stream.str();
     }
